errors are caught when PUTting and displayed with the data
fixes SPOT-769
refs SPOT-751

+++ views/edit_space.py
@@ -25,12 +25,12 @@ def edit_space(request, spot_id):
         if 'q_id' in space_datum:
             q_obj = QueuedSpace.objects.get(pk=int(request.POST["q_id"]))
             if json.loads(q_obj.json) == json.loads(cleaned_space_datum):
-                data.update({'status': q_obj.status})
+                data.update({'status': q_obj.status, 'errors': q_obj.errors})
             else:
-                data.update({'status': 'updated'})
+                data.update({'status': 'updated', 'errors': '{}'})
             form = QueueForm(data, instance=q_obj)
         else:
-            data.update({'status': 'updated'})
+            data.update({'status': 'updated', 'errors': '{}'})
             form = QueueForm(data)
         if form.is_valid():
             queued = form.save(commit=False)
@@ -41,31 +41,35 @@ def edit_space(request, spot_id):
             else:
                 queued.space_etag = space_datum['space_etag']
                 queued.space_last_modified = space_datum['space_last_modified']
-            try:
-                approved = QueuedSpace.objects.get(space_id=spot_id).approved_by
-            except:
-                approved = None
-            if 'changed' in space_datum:
+            if 'changed' in space_datum and not json.loads(queued.errors):
                 if space_datum['changed'] == 'approved':
                     queued.status = 'approved'
                     queued.approved_by = request.user
                 elif space_datum['changed'] == 'publish':
-                    spot = QueuedSpace.objects.get(space_id=spot_id)
-                    response = upload_data(request, [{'data': spot.json, 'id': spot_id}])  # PUT if spot_id, POST if not spot_id
+                    response = upload_data(request, [{'data': queued.json, 'id': spot_id}])  # PUT if spot_id, POST if not spot_id
                     if not response['failure_descs']:  # if there are no failures
                         QueuedSpace.objects.get(space_id=spot_id).delete()
                         return HttpResponseRedirect('/')
                     else:
+                        errors = {}
                         for failure in response['failure_descs']:
-                            failure
-                            #print '\n'
-                            #print 'spot: ' + failure['fname'] + '\n' + 'location: ' + failure['flocation']
-                            #for reason in failure['freason']:
-                            #    print 'reason: ' + reason
-                            #print '\n'
-                        #blow up and throws some phat errors, yo!
-            elif approved:
-                queued.approved_by = approved
+                            errors.update({failure['flocation']: []})
+                            for reason in failure['freason']:
+                                errors[failure['flocation']].append(reason)
+                        queued.errors = json.dumps(errors)
+                        queued.status = 'updated'
+                        queued.approved_by = None
+                        queued.save()
+                        url = '/space/%s' % space_datum['id']
+                        return HttpResponseRedirect(url)
+            elif 'changed' in space_datum and queued.errors:
+                url = '/space/%s' % space_datum['id']
+                return HttpResponseRedirect(url)
+            else:
+                if data['status'] == 'approved':
+                    queued.approved_by = QueuedSpace.objects.get(space_id=spot_id).approved_by
+                else:
+                    queued.approved_by = None
             queued.save()
         else:
             #TODO: do something appropriate if the form isn't valid
@@ -91,6 +95,7 @@ def edit_space(request, spot_id):
             status = spot.status
             space_etag = spot.space_etag
             space_last_modified = spot.space_last_modified
+            errors = json.loads(spot.errors)
             spot = json.loads(spot.json)
         except:
             spot_url = "%s/api/v1/spot/%s" % (settings.SS_WEB_SERVER_HOST, spot_id)
@@ -103,6 +108,7 @@ def edit_space(request, spot_id):
             space_etag = resp['etag']
             last_modified = None
             space_last_modified = spot["last_modified"]
+            errors = None
 
         args = {
             "spot_id": spot_id,
@@ -115,6 +121,7 @@ def edit_space(request, spot_id):
             "status": status,
             "space_etag": space_etag,
             "space_last_modified": space_last_modified,
+            "errors": errors,
         }
 
         context = RequestContext(request, {})
